requirements.txt
flask
requests
uuid
cryptography
pycrypto
ecdsa
grpcio
grpcio-tools
protobuf
pyOpenSSL